+++ changelog.md
@@ -1,6 +1,8 @@
 latest
  - Amplitude threshold condense selected action, enter an upper and lower amplitude value (between 0 and 1), the sample will be modified to only include amplitudes within the given range, gaps between the audio is removed condensing the sample size.
  - Ctrl + Click on Filename column header to sort the list by numeric filename content only.
+ - Improved drop file order consistency.
+ - Allow dropping mp3 files.
 
 2024.05.14 1.4.14
  - Added basic error catch in the joinAll method to help surface chain build errors.